requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
openai==1.85.0
faiss-cpu==1.11.0
transformers==4.52.4
torch==2.7.1
sentence-transformers==4.1.0
python-dotenv==1.1.0